+++ STM32F0-Utilities/STM32F0-Utilities/src/STM32F0_ADC.c
@@ -33,7 +33,7 @@ void init_ADC(int resolution) {
 	}
 }
 
-int analogRead(IOPin_TypeDef* iopin) {
+uint16_t analogRead(IOPin_TypeDef* iopin) {
 	// Read an analog value from a pin
 	ADC1->CHSELR &= 0xFFF80000; // Reset the channel selection
 	if ((iopin->port == GPIOA) && (iopin->pin == 0)) { // PA0
@@ -91,18 +91,18 @@ int analogRead(IOPin_TypeDef* iopin) {
 
 	ADC1->CR |= (1 << 2); // Set the ADSTART bit to start a conversion
 	while ((ADC1->ISR & (1 << 2)) == 0); // Wait for the conversion to complete
-	return ADC1->DR; // Return the ADC data
+	return (uint16_t)ADC1->DR; // Return the ADC data
 
 }
 
-int analogReadChannel(int channel) {
+uint16_t analogReadChannel(int channel) {
 	// Read an analog value from a specific ADC channel
 	ADC1->CHSELR &= 0xFFF80000; // Reset the channel selection
 	if ((channel >= 0) && (channel <= 18)) { // Check that the channel requested is valid
 		ADC1->CHSELR |= (1 << channel); // Select the channel
 		ADC1->CR |= (1 << 2); // Set the ADSTART bit to start a conversion
 		while ((ADC1->ISR & (1 << 2)) == 0); // Wait for the conversion to complete
-		return ADC1->DR; // Return the ADC data
+		return (uint16_t)ADC1->DR; // Return the ADC data
 	}
 	return 0; // Invalid channel selection, do not read ADC
 }
